Detach OAuth functions from API class
