+++ src/API.ts
@@ -16,7 +16,6 @@
  *
  * */
 
-import Bridge from './Bridge.js';
 import * as JSON from './JSON/index.js';
 import REST from './REST.js';
 
@@ -1034,12 +1033,6 @@ export namespace API {
         api_version?: number;
     }
 
-    export interface OAuthApp {
-        id: string;
-        client_id: string;
-        client_secret: string;
-    }
-
     export interface QueryParams extends REST.ParamRecord {
         /**
          * Maximum number of results to return. Server defaults to 20 statuses.
@@ -1098,138 +1091,6 @@ export namespace API {
          * Do not get statuses with any of these tags.
          */
         'none[]'?: Array<string>;
-    }
-
-    /* *
-     *
-     *  Functions
-     *
-     * */
-
-    /**
-     * Creates an application in a Mastodon account.
-     *
-     * @memberof API
-     *
-     * @param apiURL
-     * API URL of the Mastodon server.
-     *
-     * @param [clientName]
-     * Public name of the application.
-     *
-     * @param [redirectURI]
-     * OAuth URI.
-     *
-     * @param [scopes]
-     * Application permissions to grant.
-     *
-     * @param [website]
-     * Public website of the application.
-     *
-     * @return
-     * Promise with an object of applications `id`, `client_id` and
-     * `client_secret`.
-     */
-    export async function createOAuthApp (
-        apiURL: string,
-        clientName = 'mastodon-node',
-        redirectURI = 'urn:ietf:wg:oauth:2.0:oob',
-        scopes = 'read write follow',
-        website?: string
-    ): Promise<API.OAuthApp> {
-        const body: ( FormData | undefined ) = new Bridge.FormData();
-
-        body.append( 'client_name', clientName );
-        body.append( 'redirect_uris', redirectURI );
-        body.append( 'scopes', scopes );
-
-        if ( website ) {
-            body.append( 'website', website );
-        }
-
-        const response = await Bridge.fetch(
-            `${apiURL}apps`,
-            {
-                body,
-                method: 'POST'
-            }
-        );
-
-        return await response.json();
-    }
-
-    /**
-     * Gets the access token for the application.
-     *
-     * @memberof API
-     *
-     * @requires oauth
-     */
-    export async function getAccessToken (
-        baseURL: string,
-        clientId: string,
-        clientSecret: string,
-        authorizationCode: string,
-        redirectUri = 'urn:ietf:wg:oauth:2.0:oob'
-    ): Promise<string> {
-        const OAuth2 = ( await import( 'oauth' ) ).OAuth2;
-
-        const oauth = new OAuth2(
-            clientId,
-            clientSecret,
-            baseURL,
-            undefined,
-            '/oauth/token'
-        );
-
-        return new Promise( ( resolve, reject ) => {
-            oauth.getOAuthAccessToken(
-                authorizationCode,
-                {
-                    grant_type: 'authorization_code',
-                    redirect_uri: redirectUri
-                },
-                ( err, accessToken ) => {
-                    if ( err ) {
-                        reject( err )
-                        return
-                    }
-                    resolve( accessToken )
-                }
-            );
-        } );
-    }
-
-    /**
-     * Creates an authorization url for users to authorize the application.
-     *
-     * @memberof API
-     *
-     * @requires oauth
-     */
-    export async function getAuthorizationUrl (
-        baseURL: string,
-        clientId: string,
-        clientSecret: string,
-        redirectURI = 'urn:ietf:wg:oauth:2.0:oob',
-        scope = 'read write follow'
-    ): Promise<string> {
-        const OAuth2 = ( await import( 'oauth' ) ).OAuth2;
-
-        const oauth = new OAuth2(
-            clientId,
-            clientSecret,
-            baseURL,
-            undefined,
-            '/oauth/token'
-        );
-
-        return oauth.getAuthorizeUrl( {
-            redirect_uri: redirectURI,
-            response_type: 'code',
-            client_id: clientId,
-            scope
-        } );
     }
 
 }
